Separate Update Implementation into specific components
Currently everything is implemented in the mojo that might mixes up
different concerns.

This now separates the update login into dedicated components so they
are easier to understand and use.
